refactor: 추상클래스 중복 메서드 구현
